#52 Removing and organizing code for better enablement
