+++ src/app/api/groq/threads/[id]/runs/route.ts
@@ -1,10 +1,37 @@
 import { PrismaClient } from '@prisma/client';
 import { NextRequest, NextResponse } from 'next/server';
-import { getGroqObjectForAssistant } from '@/app/api/utils';
 import { ulid } from 'ulidx';
 import { createMessage } from '@/app/api/utils/messages';
+import { Groq } from 'groq-sdk';
 
 const prisma = new PrismaClient();
+
+export const getGroqObjectForAssistant = async (
+  req: NextRequest,
+  prisma: PrismaClient
+) => {
+  let assistantId = req.headers.get('X-Assistant-Id');
+
+  // @ts-ignore
+  let assistant = await prisma.assistant.findFirst({
+    where: {
+      id: assistantId ? assistantId : undefined,
+    },
+    select: {
+      organization: true,
+      modelId: true,
+      object: true,
+    },
+  });
+
+  if (!assistant) {
+    throw new Error('Assistant does not exist');
+  }
+
+  return new Groq({
+    apiKey: assistant?.organization?.groqCloudApiKey,
+  });
+};
 
 const getId = (req: Request) => {
   const url = new URL(req.url);
